HeadingExtractor/requirements.txt
pdfminer.six
numpy
pandas
sentence-transformers
torch
